fix(Runtime): added epitech headers

+++ common/Runtime/ECS/Entity.hpp
@@ -5,6 +5,6 @@
 namespace RType::Runtime::ECS
 {
 
-    using Entity = std::uint64_t;
+    using Entity = std::uint32_t;
 
 } // namespace RType::Runtime::ECS
